Data.Rhythm.Christoffel

Description

Christoffel words.

Documentation

christoffelWord :: Bool -> Integer -> Integer -> Maybe Int -> [Integer]

Generate the upper or lower Christoffel word for a given slope with a given number of terms.

>>> christoffelWord False 3 7 Nothing
[0,0,0,1,0,0,1,0,0,1]

>>> christoffelWord True 3 7 Nothing
[1,0,0,1,0,0,1,0,0,0]

>>> christoffelWord True 3 7 (Just 20)
[1,0,0,1,0,0,1,0,0,0,1,0,0,1,0,0,1,0,0,0]